gcp_cloud_functions/thumbnails/requirements.txt
google-cloud-storage
Pillow